feat: generalize utils config functions to be used by mainnet, testnet and sandbox chains
